Standarlize and minor fixes

+++ ft_bzero.c
@@ -6,13 +6,15 @@
 /*   By: zchoo <zchoo@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/11/20 14:20:44 by zchoo             #+#    #+#             */
-/*   Updated: 2025/11/20 14:39:14 by zchoo            ###   ########.fr       */
+/*   Updated: 2025/11/27 13:44:39 by zchoo            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	ft_bzero(void *s, unsigned int n)
+#include "libft.h"
+
+void	ft_bzero(void *s, size_t n)
 {
-	unsigned int	i;
+	size_t			i;
 	unsigned char	*bytes;
 
 	i = 0;
