+++ ft_memset.c
@@ -6,13 +6,15 @@
 /*   By: zchoo <zchoo@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/11/20 14:58:47 by zchoo             #+#    #+#             */
-/*   Updated: 2025/11/22 16:05:23 by zchoo            ###   ########.fr       */
+/*   Updated: 2025/11/27 13:46:43 by zchoo            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	*ft_memset(void *s, int c, unsigned int n)
+#include "libft.h"
+
+void	*ft_memset(void *s, int c, size_t n)
 {
-	unsigned int	i;
+	size_t			i;
 	unsigned char	*bytes;
 
 	i = 0;
